backend/requirements.txt
# FastAPI and web server
fastapi==0.124.1
uvicorn[standard]==0.24.0
python-multipart>=0.0.9,<0.0.13


# Database
alembic==1.12.1
sqlalchemy[asyncio]==2.0.23
psycopg[binary]==3.3.2
asyncpg==0.29.0

# OpenAI and AI
openai>=2.8.0  # OpenAI Python SDK (version 2.8+ required by openai-agents)
openai-agents>=0.4.0  # OpenAI Agents SDK for multi-agent workflows

# Vector database
qdrant-client==1.7.1

# Authentication
# Note: Better Auth is primarily a Node.js library
# We'll use standard Python auth libraries instead
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Validation and data models
pydantic>=2.12.3  # Required by openai-agents>=0.4.0
pydantic-settings>=2.1.0
email-validator==2.1.0

# Utilities
bleach==6.1.0  # For input sanitization
slowapi==0.1.9  # For rate limiting
redis==5.0.1  # For rate limiting backend

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.27.0,<0.30  # For testing FastAPI
secure